[move] serviceダイアログページのディレクトリ移動
